Replace MultipleContextTest with InternalAssertionTest
